Reduce memory and time requirement for Lomb-Scargle spectral calculation.
The calculation is done only on frequencies within the bands that
will be averaged.  The number of frequencies in a band is limited
by a parameter that is initially set to 500; this can be changed
and/or made an option later, if desired.  I suspect 100 would be
adequate.

I also modified the py.test options in setup.cfg; there were a
couple that were not recognized, and others that it seemed reasonable
to change.