Finished functional aspects of FUV, added test
FUV now handles valid flag combinations, including
nodal corrections and Greenwich phase.

The output of FUV relies on broadcasting rather than
replication in some, but not all, cases where an output
is independent of the time variable.  If it turns out
that this doesn't cause problems elsewhere, then the
use of broadcasting can be extended to the remaining
cases.

The docstrings in this module remain to be redone.