Add smoke test for initial work on inference, and make it run.
This changeset includes quite a range of modifications; ideally
it might have been broken up into smaller pieces.

The data from ut_constants is now run through a new function,
convert_unicode_arrays, so that things like the list of
constituent names are object arrays of strings with trailing
whitespace removed.

The new test, test_HNL.py, is presently only a smoke test,
but it will be expanded to compare results against the
matlab-generated output that is now in tests/data.