Include more information in the return from reconstruct().
The returned Bunch now includes the kwargs and the times,
with the latter both in the original form and in matplotlib
datenums.

Added two convenience Bunch objects to the utide namespace:
hours_per_cycle and cycles_per_hour, both indexed with the
constituent name.

Updated the utide_real_data_example.ipynb notebook, and
stored it without output.