Fix Travis test failure via improved test.
When given a pure tidal sinusoid, the confidence interval calculation
can fail in a platform-specific way.  Ideally this might be fixed in the
core code, but it's not trivial to do so, and the problem occurs only in
highly artificial testing situations.  Therefore we tweak the test,
adding a small amount of noise to the test time series to make the
algorithm work normally.