:bug: Updates time input handling
Time is a very sensitive parameter for reproducing inter-software
comparable phases in tidal analysis. utide use a time convention of
gregorian datenum, where 1st day of the first year - 0001-01-01 -
corresponds to day 1. Which essentially means an epoch of '0000-12-31'.

Previously this datenum was converted using matplotlib.dates.date2num
function. This function, however, went through changes over time, which
changed its output from datenum with an epoch '0000-12-31' to the unix
epoch '1970-01-01'. This happened during transition from python2 to
python 3. In python2, the ordinal of '0001-01-01' is 1, whereas, in
python3 is it -719162. These changes also changed the matplotlib
behaviour over time. Thus the phases from the utide analysis, where
the time is converted to datenum using matplotlib.dates.date2num was not
actually referenced to greenwich time anymore.

This commit implements a convertion based on numpy datenum64 object,
which is also the basis of the pandas datetime, and completely
inter-operable between versions. The following changes are implemented for
both solving and reconstruction methods -
    1. The keyword `epoch` is now set to None as default, when
    the time array is a true datetime array. As most of data analysis
    these days uses pandas dataframe, it allows to pass pandas datetime
    array directly. The final output datenum of t follows the utide
    convention, i.e., day 1 of year 1 is equals to 1.
    2. If `epoch` is python, then the time array is considered a float
    or integer array of datenum with an epoch of '0000-31-12'. It
    retains the previous behaviour. Similarly, matlab `epoch` also
    retains its previous behaviour.
    3. additionally, tests for time normalization is updated to check the
    time conversion behaviour with pytest. All the corresponding test
    lines following this change, e.g., explicit kwarg epoch='python', is
    implemented.
    4. The notebooks are updated following this change.

This changes are tested to work interoperably between python2 and
python3. All the unittests passed in python3. The results are compared
with another software with greenwich phase output - comodo tools
(LEGOS/France). The resulting phases are now comparable between these
two softwares.